pyproject.toml
[project]
name = "text-to-sql"
version = "0.1.0"
description = "Conversational text-to-SQL system with LangGraph workflow for generating SQL queries from natural language"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Text-to-SQL Team" }
]
keywords = [
    "text-to-sql",
    "langgraph",
    "langchain",
    "sql",
    "nlp",
    "conversational-ai",
    "streamlit"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    # Core dependencies
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "typing-extensions>=4.12.0",
    
    # LangChain ecosystem
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",
    
    # Web framework
    "streamlit>=1.39.0",
    
    # Database
    "sqlalchemy>=2.0.36",
    "sqlparse>=0.5.0",
    
    # Data processing
    "pandas>=2.2.0",
    
    # Configuration
    "pyyaml>=6.0.1",
    
    # Data generation (optional, for synthetic data)
    "faker>=27.0.0",
    "plotly>=5.24.0",
]

[project.optional-dependencies]
# Database-specific drivers
postgresql = ["psycopg2-binary>=2.9.9"]
databricks = ["databricks-sql-connector>=3.0.0"]
mysql = ["pymysql>=1.1.0"]

[project.urls]
Homepage = "https://github.com/yourusername/text-to-sql"
Documentation = "https://github.com/yourusername/text-to-sql#readme"
Repository = "https://github.com/yourusername/text-to-sql"
Issues = "https://github.com/yourusername/text-to-sql/issues"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["workflows", "utils", "prompts"]
include = [
    "data/generate_synthetic_data.py",
    "data/config.json",
    "workflow_config.json",
    "app.py",
]

